+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ### Added
 
+- full logic Brigitte
 - rubygems support
 - serialization